# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sniper Duels Auto Shop API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Sniper Duels"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative (same-origin)"
    }
  ],
...
  "paths": {
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "operationId": "getMe",
        "summary": "Get current user info",
        "description": "Returns the authenticated user's profile, wallet balance, loyalty tier, and admin status. Auth is optional — returns null user if unauthenticated.",
        "security": [{ "sessionCookie": [] }, {}],
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "nullable": true,
                      "$ref": "#/components/schemas/User"
                    },
                    "isAdmin": { "type": "boolean" },
                    "discordLinked": { "type": "boolean" },
                    "discordUsername": { "type": "string", "nullable": true },
                    "walletBalance": { "type": "number", "format": "double", "description": "USD balance" },
                    "loyaltyTier": { "type": "string", "nullable": true, "description": "Current loyalty tier name" },
                    "loyaltyDiscount": { "type": "number", "format": "double", "description": "Discount multiplier (e.g. 0.05 = 5%)" },
                    "lifetimeSpend": { "type": "number", "format": "double" },
                    "canUseDiscordDiscount": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/redirect": {
      "get": {
        "tags": ["Auth"],
        "operationId": "robloxOAuthCallback",
        "summary": "Roblox OAuth callback",
        "description": "Handles the OAuth callback from Roblox. Exchanges the authorization code for user info, creates a JWT session cookie, and redirects to the site.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Authorization code from Roblox"
          },
          {
            "name": "state",
            "in": "query",
            "schema": { "type": "string" },
            "description": "CSRF state parameter"
          },
          {
            "name": "error",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Error code if OAuth failed"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to site after login"
          }
        }
      }
    },
...
    "/api/deposits": {
      "get": {
        "tags": ["Deposits"],
        "operationId": "getDeposits",
        "summary": "Get user's deposits",
        "description": "Returns all deposits for the authenticated user.",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "List of deposits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Deposit" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/admin/orders/{id}/cancel": {
      "post": {
        "tags": ["Admin"],
        "operationId": "adminCancelOrder",
        "summary": "Cancel order and refund",
        "description": "Cancels an order and refunds the purchase amount to the user's wallet.",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled and refunded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": { "$ref": "#/components/schemas/Order" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/admin/users/{id}/balance": {
      "post": {
        "tags": ["Admin"],
        "operationId": "adminAdjustBalance",
        "summary": "Adjust user balance",
        "description": "Adds to, removes from, or sets a user's wallet balance.",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "User ID (Roblox user ID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "action"],
                "properties": {
                  "amount": { "type": "number", "format": "double", "description": "Amount in USD" },
                  "action": {
                    "type": "string",
                    "enum": ["add", "remove", "set"],
                    "description": "Balance adjustment action"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Balance adjusted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "newBalance": { "type": "number", "format": "double" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/bot/orders": {
      "get": {
        "tags": ["Bot"],
        "operationId": "botGetOrders",
        "summary": "Get pending orders for bot",
        "description": "Returns orders that are ready for the delivery bot to process.",
        "security": [{ "botApiKey": [] }],
        "parameters": [
          {
            "name": "include_processing",
            "in": "query",
            "schema": { "type": "boolean", "default": false },
            "description": "Include orders already being processed"
          }
        ],
        "responses": {
          "200": {
            "description": "Pending orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Order" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "JWT session cookie obtained via Roblox OAuth. httpOnly, signed with SESSION_SECRET."
      },
      "botApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-bot-api-key",
        "description": "API key for the delivery bot."
      }
    },
    "parameters": {
      "OrderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Order ID"
      },
...
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Roblox user ID" },
          "name": { "type": "string", "description": "Roblox username" },
          "displayName": { "type": "string", "description": "Roblox display name" },
          "avatar": { "type": "string", "format": "uri", "description": "Roblox avatar thumbnail URL" }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string", "description": "Roblox user ID of buyer" },
          "itemId": { "type": "string", "nullable": true, "description": "Stock item ID (null for gem orders)" },
          "itemName": { "type": "string" },
          "quantity": { "type": "integer" },
          "totalUsd": { "type": "number", "format": "double" },
          "status": {
            "type": "string",
            "enum": ["pending", "ready", "processing", "completed", "failed", "cancelled"],
            "description": "Current order status"
          },
          "botTradeId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Deposit": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "amount": { "type": "number", "format": "double", "description": "Deposit amount in USD" },
          "status": {
            "type": "string",
            "enum": ["pending", "completed", "failed"],
            "description": "Payment status"
          },
          "checkoutUrl": { "type": "string", "format": "uri", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Not authenticated or invalid session",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions (not admin / not owner)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}